refactor-#158: move booking service to services
