+++ src/microservices/booking/booking.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { ClientsModule } from '@nestjs/microservices';
 
-import { BookingService } from './booking.service';
+import { createClientAsyncOptions } from '@utils/client';
+import { BookingService } from './services';
 import { BookingController } from './rest/booking.controller';
 import { BookingMutationResolver } from './graphql/booking-mutation.resolver';
 import { BookingQueryResolver } from './graphql/booking-query.resolver';
-import { createClientAsyncOptions } from '@utils/client';
 
 @Module({
   imports: [
